ivmsbackend/requirements.txt
asgiref==3.6.0
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
dj-database-url==1.3.0
dj-email-url==1.0.6
dj-rest-auth==3.0.0
Django==4.2
django-allauth==0.54.0
django-cache-url==3.4.4
django-cors-headers==3.14.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
environs==9.5.0
gunicorn==20.1.0
idna==3.4
importlib-resources==5.12.0
inflection==0.5.1
jsonschema==4.17.3
marshmallow==3.19.0
oauthlib==3.2.2
packaging==23.1
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
whitenoise==6.4.0
zipp==3.15.0
